handling edge cases
